feat: refactor person module structure and add ApiErrorResponse and ApiException classes
