+++ app/src/main/java/com/giraffe/cineverseapp/di/localDataSourceModule.kt
@@ -4,9 +4,9 @@ import com.giraffe.cineverseapp.data.database.CineVerseDatabase
 import com.giraffe.explore.LocalExploreDataSourceImpl
 import com.giraffe.explore.dao.ExploreSearchKeywordDao
 import com.giraffe.explore.datasource.local.LocalExploreDataSource
-import com.giraffe.person.PersonLocalDatasource
 import com.giraffe.person.PersonLocalDatasourceImp
 import com.giraffe.person.dao.PersonDao
+import com.giraffe.person.local.PersonLocalDatasource
 import org.koin.dsl.module
 
 val localDataSourceModule = module {
