+++ app/src/main/java/com/giraffe/cineverseapp/di/networkModule.kt
@@ -4,8 +4,8 @@ import com.giraffe.cineverseapp.BuildConfig
 import com.giraffe.cineverseapp.data.network.HttpClientFactory
 import com.giraffe.explore.RemoteExploreDataSourceImpl
 import com.giraffe.explore.datasource.remote.RemoteExploreDataSource
-import com.giraffe.person.PersonRemoteDataSource
 import com.giraffe.person.PersonRemoteDataSourceImp
+import com.giraffe.person.remote.PersonRemoteDataSource
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
