+++ datasource/local/person/build.gradle.kts
@@ -33,17 +33,12 @@ android {
 }
 
 dependencies {
+    implementation(project(":repository:person"))
+    implementation(libs.room.runtime)
+    implementation(libs.room.ktx)
+    annotationProcessor(libs.room.compiler)
 
-    implementation(project(":domain:person"))
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    implementation(libs.material)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    annotationProcessor(libs.room.compiler)
-    implementation(libs.room.runtime)
-    implementation(libs.room.ktx)
-    implementation(libs.koin.android)
-    implementation(libs.koin.androidx.compose)
 }
