+++ datasource/local/person/src/main/java/com/giraffe/person/PersonLocalDatasourceImp.kt
@@ -1,7 +1,8 @@
 package com.giraffe.person
 
 import com.giraffe.person.dao.PersonDao
-import com.giraffe.person.dto.PersonDto
+import com.giraffe.person.local.PersonLocalDatasource
+import com.giraffe.person.local.dto.PersonDto
 
 class PersonLocalDatasourceImp(private val dao: PersonDao) : PersonLocalDatasource {
     override suspend fun storePerson(person: PersonDto) = dao.storePerson(person)
