+++ datasource/local/person/src/main/java/com/giraffe/person/dao/PersonDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.giraffe.person.dto.PersonDto
+import com.giraffe.person.local.dto.PersonDto
 
 @Dao
 interface PersonDao {
