+++ datasource/remote/person/build.gradle.kts
@@ -35,6 +35,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":repository:person"))
     implementation(libs.bundles.ktor)
 
     implementation(libs.androidx.core.ktx)
