+++ repository/person/src/main/java/com/giraffe/person/remote/PersonRemoteDataSource.kt
@@ -1,6 +1,6 @@
-package com.giraffe.person
+package com.giraffe.person.remote
 
-import com.giraffe.person.response.SearchPersonResponse
+import com.giraffe.person.remote.response.SearchPersonResponse
 
 interface PersonRemoteDataSource {
     suspend fun searchByName(personName: String): SearchPersonResponse
