+++ repository/person/src/main/java/com/giraffe/person/remote/response/ApiErrorResponse.kt
@@ -1,0 +1,13 @@
+package com.giraffe.person.remote.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ApiErrorResponse(
+    @SerialName("status_code")
+    val statusCode: Int,
+    @SerialName("status_message")
+    val statusMessage: String,
+    val success: Boolean
+)
