+++ repository/person/src/main/java/com/giraffe/person/util/ApiException.kt
@@ -1,0 +1,3 @@
+package com.giraffe.person.util
+
+class ApiException(val code: Int) : Exception()
