--- repository/person/src/main/java/com/giraffe/person/util/ConnectionChecker.kt
@@ -1,22 +1,0 @@
-package com.giraffe.person.util
-
-import android.Manifest
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import androidx.annotation.RequiresPermission
-
-class ConnectionChecker(
-    private val context: Context
-) {
-    @RequiresPermission(Manifest.permission.ACCESS_NETWORK_STATE)
-    fun isConnected(): Boolean {
-        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-
-        val network = connectivityManager.activeNetwork ?: return false
-        val networkCapabilities = connectivityManager.getNetworkCapabilities(network) ?: return false
-
-        return networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) &&
-                networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_VALIDATED)
-    }
-}
